feat: Refactor Keycloak mappers package structure.
Moved `UserRepresentationModelMapper` and `UserToUserRepresentationMapper` to a dedicated `mappers` subpackage for better organization. Updated imports across the codebase to reflect the new package structure. This change improves maintainability and aligns with package organization best practices.
