+++ lsadf_core/src/main/java/com/lsadf/core/application/user/UserServiceImpl.java
@@ -21,8 +21,8 @@ import com.lsadf.core.infra.exceptions.AlreadyExistingUserException;
 import com.lsadf.core.infra.exceptions.http.InternalServerErrorException;
 import com.lsadf.core.infra.exceptions.http.NotFoundException;
 import com.lsadf.core.infra.web.config.keycloak.KeycloakProperties;
-import com.lsadf.core.infra.web.config.keycloak.UserRepresentationModelMapper;
-import com.lsadf.core.infra.web.config.keycloak.UserToUserRepresentationMapper;
+import com.lsadf.core.infra.web.config.keycloak.mappers.UserRepresentationModelMapper;
+import com.lsadf.core.infra.web.config.keycloak.mappers.UserToUserRepresentationMapper;
 import com.lsadf.core.infra.web.requests.user.creation.UserCreationRequest;
 import com.lsadf.core.infra.web.requests.user.creation.UserCreationRequestModelMapper;
 import com.lsadf.core.infra.web.requests.user.update.AdminUserUpdateRequest;
