+++ lsadf_core/src/test/java/com/lsadf/core/unit/config/UnitMapperMockConfiguration.java
@@ -22,8 +22,8 @@ import com.lsadf.core.infra.persistence.mappers.game.GameSaveEntityModelMapper;
 import com.lsadf.core.infra.persistence.mappers.game.StageEntityModelMapper;
 import com.lsadf.core.infra.persistence.mappers.game.inventory.InventoryEntityModelMapper;
 import com.lsadf.core.infra.persistence.mappers.game.inventory.ItemEntityModelMapper;
-import com.lsadf.core.infra.web.config.keycloak.UserRepresentationModelMapper;
-import com.lsadf.core.infra.web.config.keycloak.UserToUserRepresentationMapper;
+import com.lsadf.core.infra.web.config.keycloak.mappers.UserRepresentationModelMapper;
+import com.lsadf.core.infra.web.config.keycloak.mappers.UserToUserRepresentationMapper;
 import com.lsadf.core.infra.web.requests.game.characteristics.CharacteristicsRequestModelMapper;
 import com.lsadf.core.infra.web.requests.game.currency.CurrencyRequestModelMapper;
 import com.lsadf.core.infra.web.requests.game.stage.StageRequestModelMapper;
